requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
gunicorn==21.2.0
idna==3.6
packaging==23.2
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
